requirements.txt
Flask
openai
requests
gtts
python-dotenv
assemblyai
soundfile
numpy
gevent